Fix 64-bit type truncation warnings in GUI coordinate handling
- Change helper_rubber_config and helper_tray_config to use int for x/y/width/height coordinates
- Fix cl_slider.c to explicitly cast long GUI dimensions to int before arithmetic
- Ensure explicit type conversions when passing coordinates to GLUpdateGeometry()
- All GUI coordinates are safely int-sized (physical screen dimensions < INT_MAX)
- Add explicit casts in cl_txtDisp.c for height/width assignments
- Fix type handling in cl_txtLabel.c for xpa_centered() parameters

These changes eliminate implicit 64-to-32 bit truncation in geometry calculations
while maintaining semantic correctness - GUI coordinates are always int in X11.

+++ src/viola/cl_BCard.c
@@ -144,7 +144,7 @@ long meth_BCard_expose(VObj* self, Packet* result, int argc, Packet argv[]) {
     return meth_field_render(self, result, argc, argv);
 }
 
-long helper_BCard_get(VObj* self, Packet* result, int argc, Packet argv[], int labelID) {
+long helper_BCard_get(VObj* self, Packet* result, int argc, Packet argv[], long labelID) {
     return helper_field_get(self, result, argc, argv, labelID);
 }
 
@@ -166,7 +166,7 @@ long meth_BCard_render(VObj* self, Packet* result, int argc, Packet argv[]) {
     return 1;
 }
 
-long helper_BCard_set(VObj* self, Packet* result, int argc, Packet argv[], int labelID) {
+long helper_BCard_set(VObj* self, Packet* result, int argc, Packet argv[], long labelID) {
     return helper_field_set(self, result, argc, argv, labelID);
 }
